+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 26.1.2.1
 
 * Add fun buttons to spice up the mob farm
+* Add ability to configure mod item chest loot tables
 
 ## 26.1.2.0